% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{classify_transaction}
\alias{classify_transaction}
\title{Classify transaction description using key words}
\usage{
classify_transaction(description, transaction_coding)
}
\arguments{
\item{description}{string description of transaction}

\item{transaction_coding}{list of pattern vectors defining transaction
types}
}
\value{
Returns of named vector with transaction type and pattern that
matched
}
\description{
Classifies a transaction description type using a key word dictionary
(list) describing key words (patterns) for different transaction types
}
