% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{classify_transactions}
\alias{classify_transactions}
\title{Classify transaction descriptions using key words}
\usage{
classify_transactions(descriptions, transaction_coding)
}
\arguments{
\item{descriptions}{vector of string description of transactions}

\item{transaction_coding}{list of pattern vectors defining transaction types}
}
\value{
Returns of named vector with transaction type and pattern that
matched
}
\description{
Classifies types for transaction descriptions type using a key word
dictionary
(list) describing key words (patterns) for different transaction types
}
