% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{clean_strings}
\alias{clean_strings}
\title{Remove weird characters from strings}
\usage{
clean_strings(strings)
}
\arguments{
\item{strings}{vector of character strings}
}
\value{
Returns strings without hidden characters or non-alphanumeric
characters
}
\description{
Removes any hidden characters or non-alphanumeric characters from strings
}
