% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{generate_transaction_dates}
\alias{generate_transaction_dates}
\title{Generate random dates for transactions}
\usage{
generate_transaction_dates(
  start_date,
  end_date,
  frequency,
  random_n_per_month = 4,
  day_of_month = 1,
  day_of_week = 1
)
}
\arguments{
\item{start_date}{Date object for start date}

\item{end_date}{Date object for end date}

\item{frequency}{frequency that transaction type seen in transactions.
Expecting one of c("monthly", "weekly", "daily",
"weekdays", "random")}

\item{random_n_per_month}{if random frequency, on average how many
transactions per month. Defaults to 4.}

\item{day_of_month}{if monthly, which day of month. Defaults to 1 (first
day).}

\item{day_of_week}{if weekly, which day of week. Defaults to 1 (first
day).}
}
\value{
vector of dates for transactions
}
\description{
Given start and end date for period and frequency of transaction this
function provides dates for the transactions.
}
