% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{generate_transactions}
\alias{generate_transactions}
\title{Generate random transactions based on transaction types}
\usage{
generate_transactions(start_date, end_date, transaction_types)
}
\arguments{
\item{start_date}{Date object for start date}

\item{end_date}{Date object for end date}

\item{transaction_types}{a list with details for each transaction type
expected}
}
\value{
a dataframe of random transactions matching details of each type
of transaction
}
\description{
Given a list providing information about different transaction types
this function will generate random transactions for each type across
a set time period
}
