% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{generate_transactions_for_type}
\alias{generate_transactions_for_type}
\title{Generate random transactions of a particular type}
\usage{
generate_transactions_for_type(
  average_value,
  name,
  type,
  frequency,
  start_date,
  end_date,
  random_n_per_month = NULL,
  standard_deviation = NULL,
  day_of_month = NULL,
  day_of_week = NULL,
  patterns = NULL
)
}
\arguments{
\item{average_value}{average value for type}

\item{name}{name of transaction}

\item{type}{type of transaction ("in" (credit), or "out" (debit))}

\item{frequency}{frequency that transaction type seen in transactions.
Expecting one of c("monthly", "weekly", "daily",
"weekdays", "random")}

\item{start_date}{Date object for start date}

\item{end_date}{Date object for end date}

\item{random_n_per_month}{if random frequency, on average how many
transactions per month. Defaults to 4.}

\item{standard_deviation}{standard deviation from average value for
transaction. Defaults to 10\% of value.}

\item{day_of_month}{if monthly, which day of month. Defaults to 1 (first
day).}

\item{day_of_week}{if weekly, which day of week. Defaults to 1 (first day).}

\item{patterns}{patterns to use as transaction descriptions. Defaults to
type parameter.}
}
\value{
a dataframe of random transactions matching details of transaction
type
}
\description{
Given information about transaction type, this function will generate random
transactions across time period for that type
}
