% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{load_transaction_coding}
\alias{load_transaction_coding}
\title{Build transaction classifications from dataframe}
\usage{
load_transaction_coding(transactions_coding_df, delimiter = ";")
}
\arguments{
\item{transactions_coding_df}{dataframe of transaction Type and Patterns}

\item{delimiter}{delimiter used in pattern column to separate patterns}
}
\value{
Returns list of transaction types and patterns that define them
}
\description{
Given dataframe of transaction types and delimited patterns this
parses delimited patterns and stores results in list:
\itemize{
\item key: type
\item value: \link{patterns}
}
}
