% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{order_by_month}
\alias{order_by_month}
\title{Order dataframe by month column}
\usage{
order_by_month(by_month_data, month_column = "month")
}
\arguments{
\item{by_month_data}{dataframe with a month column}

\item{month_column}{name of month column. Defaults to "month"}
}
\value{
returns dataframe ordered by month column
}
\description{
For dataframe with month column (month formatted as \%m/\%Y) function will
order rows by this column
}
