% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_values_bar_chart}
\alias{plot_values_bar_chart}
\title{Plot horizontal bar chart of values (prices)}
\usage{
plot_values_bar_chart(values, names, ...)
}
\arguments{
\item{values}{values (length) for each bar}

\item{names}{names for each bar}

\item{...}{additional parameters to pass to barplot()}
}
\description{
Creates horizontal bar chart of values
}
