% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_values_by_month}
\alias{plot_values_by_month}
\title{Plot values by month}
\usage{
plot_values_by_month(
  data,
  month_column = "month",
  y_axis_label = "£",
  title = "Values by month",
  point_shape = 19,
  add_zero_line = TRUE,
  margins = c(5.1, 4.1, 4.1, 8),
  legend_cex = 1
)
}
\arguments{
\item{data}{dataframe with month and value columns}

\item{month_column}{column with months in. Defaults to "month"}

\item{y_axis_label}{label for Y axis. Defaults to "£"}

\item{title}{title for plot. Defaults to "Values by month"}

\item{point_shape}{shape for points on line. Defaults to 19 (circle)}

\item{add_zero_line}{boolean value indicating whether to add horizontal line
at zero on Y axis. Defaults to TRUE.}

\item{margins}{four value vector for margin sizes (bottom, left, top,
right).Defaults to c(5.1, 4.1, 4.1, 8)}

\item{legend_cex}{scaling value for legend size. Defaults to 1.}
}
\description{
Given dataframe with months in one column and values in other column(s),
this function will plot the values against the months
}
