% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotly_values_bar_chart}
\alias{plotly_values_bar_chart}
\title{Plot horizontal bar chart of values (prices)}
\usage{
plotly_values_bar_chart(values, names, title, ...)
}
\arguments{
\item{values}{values (length) for each bar}

\item{names}{names for each bar}

\item{title}{title for plot}
}
\description{
Creates horizontal bar chart of values using plotly
}
