% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotly_values_by_month}
\alias{plotly_values_by_month}
\title{Plot values by month - interactive}
\usage{
plotly_values_by_month(
  data,
  month_column = "month",
  y_axis_label = "£",
  title = "Values by month"
)
}
\arguments{
\item{data}{dataframe with month and value columns}

\item{month_column}{column with months in. Defaults to "month"}

\item{y_axis_label}{label for Y axis. Defaults to "£"}

\item{title}{title for plot. Defaults to "Values by month"}
}
\description{
Given dataframe with months in one column and values in other column(s),
this function will plot the values against the months using plotly
}
