% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing_functions.R
\name{summarise_unclassified}
\alias{summarise_unclassified}
\title{Finds and summarises transactions that couldn't be classified}
\usage{
summarise_unclassified(transactions, in_column, out_column, description_column)
}
\arguments{
\item{transactions}{dataframe containing transaction details}

\item{in_column}{name of column in transactions dataframe containing
money going in (credit)}

\item{out_column}{name of column in transactions dataframe containing
money going out (debit)}

\item{description_column}{column containing transaction description}
}
\value{
Returns dataframe of unique transaction description, their cost and
number of appearances
}
\description{
Given transactions dataframe with "Type" column indicating whether
classified. This function finds unclassified transactions, sums and
counts them.
Note, should have run classify_transactions on transactions dataframe
before using this function!
}
