% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data_functions.R
\name{write_transaction_types}
\alias{write_transaction_types}
\title{Write transaction type info to file}
\usage{
write_transaction_types(file_path, transaction_types)
}
\arguments{
\item{file_path}{path and name of file to write transaction types info to}

\item{transaction_types}{a list with details for each transaction type
expected}
}
\description{
Write transaction type info to file
}
